+++ src/Schema/Postgres/Grammar.php
@@ -29,9 +29,9 @@ class Grammar extends PostgresGrammar
 
     protected function typeNumeric(Fluent $column): string
     {
-        $type      = NumericType::TYPE_NAME;
+        $type = NumericType::TYPE_NAME;
         $precision = $column->get('precision');
-        $scale     = $column->get('scale');
+        $scale = $column->get('scale');
 
         if ($precision) {
             return "${type}({$precision}," . ($scale ? ", {$scale}" : '') . ')';
@@ -43,6 +43,16 @@ class Grammar extends PostgresGrammar
     protected function typeDateRange(Fluent $column): string
     {
         return 'daterange';
+    }
+
+    protected function typeUuidArray(Fluent $column): string
+    {
+        return 'uuid[]';
+    }
+
+    protected function typeIntArray(Fluent $column): string
+    {
+        return 'int[]';
     }
 
     protected function typeTsrange(Fluent $column): string